pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ue5_asset_gen"
version = "0.1.0"
description = "Using Large Landuage Models generate 3D assest that can be imported into the Unreal Engine"
authors = [
    "rajiv Mandya Nagaraju"
]

[tool.poetry.dependencies]
python = "^3.12"
torch = "*"
torchvision = "*"
numpy = "*"
requests = "*"
python-dotenv = "*"
open3d = "*"
loguru = "*"

[[tool.poetry.packages]]
include = "src"